Make request-context middleware share a single domain.
There are scenarios in which independent modules may make use of request-context to stash value read from a request.  If the modules aren't aware of each other, you can end up with multiple domains, each with their own `__$cntxt__` property.  The simplest solution is to make the middleware check to see if the currently-active domain _already_ has a `__$cntxt__` property, and if so, simply use it as-is, trusting each module to choose distinct namespace values.